Improve total zaps amount calculation in notifications list
