+++ app/src/main/kotlin/net/primal/android/notifications/list/ui/NotificationListItem.kt
@@ -18,6 +18,7 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.painter.Painter
 import androidx.compose.ui.platform.LocalUriHandler
 import androidx.compose.ui.res.pluralStringResource
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import net.primal.android.R
 import net.primal.android.core.compose.AvatarThumbnailsRow
@@ -25,6 +26,8 @@ import net.primal.android.core.compose.NostrUserText
 import net.primal.android.core.compose.feed.FeedPostContent
 import net.primal.android.core.compose.feed.FeedPostStatsRow
 import net.primal.android.core.compose.feed.model.FeedPostAction
+import net.primal.android.core.compose.feed.model.FeedPostUi
+import net.primal.android.core.compose.notifications.toImagePainter
 import net.primal.android.core.ext.openUriSafely
 import net.primal.android.core.utils.shortened
 import net.primal.android.notifications.domain.NotificationType
@@ -35,8 +38,95 @@ import net.primal.android.theme.colors.ReplyLightColor
 import net.primal.android.theme.colors.RepostColor
 import net.primal.android.theme.colors.ZapColor
 
+
 @Composable
 fun NotificationListItem(
+    notifications: List<NotificationUi>,
+    type: NotificationType,
+    walletConnected: Boolean,
+    onProfileClick: (String) -> Unit,
+    onNoteClick: (String) -> Unit,
+    onHashtagClick: (String) -> Unit,
+    onReplyClick: (String) -> Unit,
+    onPostLikeClick: (FeedPostUi) -> Unit,
+    onRepostClick: (FeedPostUi) -> Unit,
+    onDefaultZapClick: (FeedPostUi) -> Unit,
+    onZapOptionsClick: (FeedPostUi) -> Unit,
+    onWalletUnavailable: () -> Unit,
+) {
+
+    notifications.map { it.actionUserSatsZapped }
+
+    val activeUsersTotalSatsZapped = notifications
+        .mapNotNull { it.actionUserSatsZapped }
+        .sum()
+        .takeIf { it > 0 }
+
+    val postTotalSatsZapped = notifications
+        .firstOrNull { it.actionPost?.stats != null }
+        ?.actionPost
+        ?.stats
+        ?.satsZapped
+        .takeIf { it != null && it > 0 }
+
+    val postData = notifications.first().actionPost
+
+    NotificationListItem(
+        notifications = notifications,
+        imagePainter = type.toImagePainter(),
+        suffixText = type.toSuffixText(
+            usersZappedCount = notifications.size,
+            totalSatsZapped = when (notifications.size) {
+                1 -> activeUsersTotalSatsZapped?.shortened()
+                else -> postTotalSatsZapped?.shortened()
+            }
+        ),
+        onProfileClick = onProfileClick,
+        onPostClick = onNoteClick,
+        onHashtagClick = onHashtagClick,
+        onPostAction = { postAction ->
+            when (postAction) {
+                FeedPostAction.Reply -> {
+                    postData?.postId?.let(onReplyClick)
+                }
+
+                FeedPostAction.Zap -> {
+                    if (walletConnected) {
+                        postData?.let { postData ->
+                            onDefaultZapClick(postData)
+                        }
+                    } else {
+                        onWalletUnavailable()
+                    }
+                }
+
+                FeedPostAction.Like -> {
+                    postData?.let(onPostLikeClick)
+                }
+
+                FeedPostAction.Repost -> {
+                    postData?.let(onRepostClick)
+                }
+            }
+        },
+        onPostLongPressAction = { postAction ->
+            when (postAction) {
+                FeedPostAction.Zap -> {
+                    if (walletConnected) {
+                        postData?.let(onZapOptionsClick)
+                    } else {
+                        onWalletUnavailable()
+                    }
+                }
+
+                else -> Unit
+            }
+        },
+    )
+}
+
+@Composable
+private fun NotificationListItem(
     notifications: List<NotificationUi>,
     imagePainter: Painter,
     suffixText: String,
@@ -153,7 +243,7 @@ fun NotificationListItem(
                         onProfileClick = onProfileClick,
                         onPostClick = onPostClick,
                         onUrlClick = { localUriHandler.openUriSafely(it) },
-                        onHashtagClick = {onHashtagClick(it) },
+                        onHashtagClick = { onHashtagClick(it) },
                     )
 
                     FeedPostStatsRow(
@@ -171,4 +261,69 @@ fun NotificationListItem(
             }
         }
     }
+}
+
+@Composable
+private fun NotificationType.toSuffixText(
+    usersZappedCount: Int = 0,
+    totalSatsZapped: String? = null,
+): String = when (this) {
+    NotificationType.NEW_USER_FOLLOWED_YOU -> stringResource(id = R.string.notification_list_item_followed_you)
+
+    NotificationType.YOUR_POST_WAS_ZAPPED -> when (totalSatsZapped) {
+        null -> stringResource(id = R.string.notification_list_item_zapped_your_post)
+        else -> stringResource(
+            id = R.string.notification_list_item_zapped_your_post_for_total_amount,
+            totalSatsZapped
+        )
+    }
+
+    NotificationType.YOUR_POST_WAS_LIKED -> stringResource(id = R.string.notification_list_item_liked_your_post)
+    NotificationType.YOUR_POST_WAS_REPOSTED -> stringResource(id = R.string.notification_list_item_reposted_your_post)
+    NotificationType.YOUR_POST_WAS_REPLIED_TO -> stringResource(id = R.string.notification_list_item_replied_to_your_post)
+
+    NotificationType.YOU_WERE_MENTIONED_IN_POST -> stringResource(id = R.string.notification_list_item_mentioned_you_in_post)
+    NotificationType.YOUR_POST_WAS_MENTIONED_IN_POST -> stringResource(id = R.string.notification_list_item_mentioned_your_post)
+
+    NotificationType.POST_YOU_WERE_MENTIONED_IN_WAS_ZAPPED -> when (totalSatsZapped) {
+        null -> stringResource(id = R.string.notification_list_item_post_you_were_mentioned_in_was_zapped)
+        else -> when (usersZappedCount) {
+            1 -> stringResource(
+                id = R.string.notification_list_item_post_you_were_mentioned_in_was_zapped_for,
+                totalSatsZapped
+            )
+
+            in 2..Int.MAX_VALUE -> stringResource(
+                id = R.string.notification_list_item_post_you_were_mentioned_in_was_zapped_for_total_amount,
+                totalSatsZapped
+            )
+
+            else -> stringResource(id = R.string.notification_list_item_post_you_were_mentioned_in_was_zapped)
+        }
+    }
+
+    NotificationType.POST_YOU_WERE_MENTIONED_IN_WAS_LIKED -> stringResource(id = R.string.notification_list_item_post_you_were_mentioned_in_was_liked)
+    NotificationType.POST_YOU_WERE_MENTIONED_IN_WAS_REPOSTED -> stringResource(id = R.string.notification_list_item_post_you_were_mentioned_in_was_reposted)
+    NotificationType.POST_YOU_WERE_MENTIONED_IN_WAS_REPLIED_TO -> stringResource(id = R.string.notification_list_item_post_you_were_mentioned_in_was_replied_to)
+
+    NotificationType.POST_YOUR_POST_WAS_MENTIONED_IN_WAS_ZAPPED -> when (totalSatsZapped) {
+        null -> stringResource(id = R.string.notification_list_item_post_where_you_post_was_mentioned_was_zapped)
+        else -> when (usersZappedCount) {
+            1 -> stringResource(
+                id = R.string.notification_list_item_post_where_you_post_was_mentioned_was_zapped_for,
+                totalSatsZapped
+            )
+
+            in 2..Int.MAX_VALUE -> stringResource(
+                id = R.string.notification_list_item_post_where_you_post_was_mentioned_was_zapped_for_total_amount,
+                totalSatsZapped
+            )
+
+            else -> stringResource(id = R.string.notification_list_item_post_where_you_post_was_mentioned_was_zapped)
+        }
+    }
+
+    NotificationType.POST_YOUR_POST_WAS_MENTIONED_IN_WAS_LIKED -> stringResource(id = R.string.notification_list_item_post_where_you_post_was_mentioned_was_liked)
+    NotificationType.POST_YOUR_POST_WAS_MENTIONED_IN_WAS_REPOSTED -> stringResource(id = R.string.notification_list_item_post_where_you_post_was_mentioned_was_reposted)
+    NotificationType.POST_YOUR_POST_WAS_MENTIONED_IN_WAS_REPLIED_TO -> stringResource(id = R.string.notification_list_item_post_where_you_post_was_mentioned_was_replied_to)
 }
